store diff mode into settings
